Extract OnePointsBuilder utility (INTERNAL)
